Contract list now scales with spawned vessels (#4172)
## About The Pull Request

As it says in the title, outpost contracts will now scale per vessel
in-sector by an addition of 5, with some coding help from Zimon (I did
something incredibly stupid)

## Why It's Good For The Game

You remember those rounds you had, where you had to share 15 contracts
with close to 10 or more vessels in-round? This makes it so you won't
have worry about having to share 15 contracts for a round with a high
vessel count.

## Changelog

:cl:
balance: The outpost contract limit will now scale with the amount of
vessels in-round
/:cl: