S-700i Corona-class Touring Vessel (Re-do) (#4151)
## About The Pull Request

(PR'd on behalf of Moffball (again))
Adds the Corona-class ship, a Cybersun marketed vessel for the civilian
market. This ship is primarily geared for consumer comfort, tourism, and
exploration across the wilds of the frontier.

"Designed by Cybersun Industries with production outsourced to ISF
Spacecraft, the S-700i Corona-class Touring Vessel is Cybersun's premier
solution to eco-friendly travel. Boasting a powerful clean energy solar
array down it's sleek profile, the Corona is not just a mode of
transportation; but a statement of opulence and eco-mindfulness in an
ever-boorish Frontier."


![image](https://github.com/user-attachments/assets/0c4f8435-872b-431b-abce-7ac730159158)

<details>
<summary>StrongDMM View</summary>

![2025-02-07 17 10
55](https://github.com/user-attachments/assets/1696c7cf-feaf-4552-870e-1732942e56a9)
![2025-02-07 17 07
49](https://github.com/user-attachments/assets/fbc6178c-9800-489d-ab7d-b5cfde5a7891)
</details>

## Why It's Good For The Game

Cybersun manufactured ships are currently in scarce supply, and an
independent ship to act as the counterpart to the Dwayne (NT
manufactured indie) fills in a fun niche.

## Changelog

:cl:MemeSnorfer and Moffball
add: Corona-class Touring Vessel
/:cl: