Helmet lights and goggles consistency (#4266)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
More helmets can use helmet lights. Helmet lights can be attached and
detatched without the use of a screwdriver. Helmets can now take NGR
factional goggles and mesons, for sake of consistency.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Noticed that for a long time, the only people using helmet lights were
Inteq, and it turned out to be, because their helmets were some of the
only ones which **could** do it. This aims to remedy that by increasing
flashlight compatability to include all the X11 and M10 variants, as
well as a few others. As for the removal of the need for a screwdriver,
it didn't quite make sense that it could be attached without a tool but
to remove it, you suddenly needed a tool. Overall it's just more
convenient and logical. Lastly, for some reason helmets couldn't take
NGR or Hardliner goggles? I assume it was oversight, so now our brave
Gorlex warriors can carry their goggles on their helmet like the rest of
the helmeted folks on the frontier.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
balance: Helmet lights attach more logically
fix: NGR and Hardliner goggles can fit on helmets
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: DIB-DOG <DaneOfThieves@gmail.com>