Explorer webbing holds O2 and binocs (#4263)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This allows the explorer webbing to hold and carry emergency oxygen
tanks and binoculars
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Played a few rounds as an asteroid miner, loading up my equipment before
an excursion and I found that I was unable to fit my binoculars into my
explorer webbing. Seems to me like binoculars are a pretty reasonable
part of exploration, so this allows them to be carried. As for the
emergency O2 tanks, it seems thematic to be able to carry them along in
your rigging as well for a particularly long trip of *exploring*.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
balance: Explorer webbing can now carry emergency O2 tanks and
binoculars.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: DIB-DOG <DaneOfThieves@gmail.com>
Co-authored-by: Changelogs <action@github.com>