Adds Mil variants of Ringneck and Cobra to cargo. (#4247)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Ringneck - 1250, 350 more (including civ discount). It's smaller, should
be fine.
Cobra - 2800, 100 more, which is not even full discount difference,
because they are literally reskins of each other. Price should change
when Mil and Civ variants actually differ.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Boomslang and Viper have both civ and mil variants in cargo. These did
not for some reason.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added Military Ringneck and Cobra variants to cargo.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Theos <theubernyan@gmail.com>