Adds a Hardliners stamp (#4321)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
It next to its main inspirations- it's meant to be a mix of NGR (shared
history) and Cybersun (the employer of the groups that would even get a
stamp). The pickaxe itself is from a thingie [posted in
loredev](https://discord.com/channels/837744059291533392/837756616304361503/1344918603786485802)
a lil bit ago.

![obraz](https://github.com/user-attachments/assets/ec745911-0537-4773-8eff-c0c77268cb1d)

I've been told that it would need a redesign whenever Cybersun is
touched. I'm fine with that and really don't mind.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Banshee starts with a CS:VS stamp for some reason. It really should not.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added a Hardliners stamp.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->