CMM Atlas Tweaks (#4207)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Slightly re-maps the Atlas-class Light Armored Cruiser's armory and
re-organizes it's contents to be more organized and friendly to the
eyes. Additionally increases starting funds on the Atlas to 4000 as
opposed to the previous 1000.

The armory change in question:
![Screenshot 2025-03-06
225224](https://github.com/user-attachments/assets/07e01429-cc26-4587-bf0d-18a9b0164192)

The dorm adjustments:
![Screenshot 2025-03-06
225234](https://github.com/user-attachments/assets/fa0d536c-1c47-4344-9498-2baecc4ee6f5)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Compared to similiar ships of it's class, the Atlas class is
simultaneously underfunded, and under-equipped. This change brings their
funding at least up to par with a similar ship, but still leaves them
behind materially. This lends itself well to the scrappier militia look
of the CLIP Minutemen while also allowing them to fulfill their function
without having to grind missions round start just to be functional.
Additionally, the armory was just terrible to work out of, this should
alleviate the issue, making the space more comfortable and better
organized.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Atlas Armory has been redesigned.
balance: Atlas starting funds increased to 4000
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->