Mech Paths and 515 EX kit (#4325)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Fixes the path on the loaded MP charger

Kills a few instances of Gygax in descs and replaces them with 501p

Adds a 515 EX kit for Syndicate factions in cargo.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

I forgot to add the 515 EX kit in #3918. It's nice to be able to fly
your faction colours I think.

Also bug fixes are good.



## Changelog

:cl:
add: 515 Conversion kit
fix: Loaded MP charger path
spellcheck: Removed a few instances of Gygax
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->