ship orbit menu improvments (#4249)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
ship helms in orbit menu now display there ship name first, and crew
size in the tool tip.
move adding a helm as "intresting" to connect to shuttle which means it
wont include ruin mapped shit.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
this sucks ass 

![image](https://github.com/user-attachments/assets/35798299-da6a-41ec-9b72-0ed4939392eb)
better

![image](https://github.com/user-attachments/assets/f4a2f7a0-1128-4b09-a68d-8407dad3e972)

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Helms in the orbit menu now use the ship's name as its full name
fix: Random helms not connected to a ship don't appear in the orbit menu
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->