EHF Point Beacons (Planet Preservation) (#4246)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR adds beacons for planetary use, to preserve, rename, and send a
distress beacon from planets. The beacon needs to be secured and power
similarly to a floodlight to be used. An inserted ID is required to send
a distress beacon, and an ID from an officer is needed to rename a
planet.

Also added: IDs now have a boolean if it belongs to an officier job, and
overmap objects now have a proc to change if it's preserved. Also,
overmap tokens now properly show if the planet is being preserved.


![image](https://github.com/user-attachments/assets/611e82c5-0425-4c54-aeed-ab849cd6cdf9)


![image](https://github.com/user-attachments/assets/84ac12b9-a018-47a6-887e-6d9cd2e58db1)


![image](https://github.com/user-attachments/assets/8ea9069b-eef4-4945-9893-cd3187f9266c)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Allows people to do more on planets than just mining and ruin clearing.
Setting up outposts, colonies, or any other semi-permanent structure,
while previously possible was difficult because of planets being
unloaded when nobody is on them. Meaning crews would either need to have
someone stay behind, or prepare everything they may need before landing.
This PR lets crews preserve planets after some initial setup (requiring
some amount of power, though not neccessarily an APC), as well as
additional features benefitting construction on planets.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: EHF point beacons, which can be used to preserve planets
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->