Adds an angle grinder to the Nail (#4269)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds an angle grinding pack, bowman headset, and a pair of welding
goggles to the Nail-class boarding shuttle
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
There had been talk in months past about offering the Crying Sun some
ability to have some form of breaching tool but at the time they were
not yet coded or implemented. Now that they're done, and have been in
the game for some time, this is adds one such tool. There was also an
un-used crate just begging to have a purpose so it worked out pretty
nice.

As the Nail-class is intended for boarding, among other things, it seems
reasonable that it would have some tool by which to facilitate this
purpose. It was debated between offering a plasma cutter or breaching
explosives but ultimately decided that the angle grinder pack would be
the best option. requiring more work and prep-time to utilize.


![Screenshot 2025-02-25
150800](https://github.com/user-attachments/assets/eebbce58-7804-4bd7-a2af-eee0077a14ee)
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added a grinder and PPE to the Nail-class boarding shuttle
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Changelogs <action@github.com>