Adds Azuilhauz variants for Vox and Kepori (#4393)
## About The Pull Request

Adds Azuilhauz variants for Vox and Kepori.
Thanks to Thgvr for making the Kepori cape's base sprite for me and to
Cloudbreak for helping me figure out the code.

![image](https://github.com/user-attachments/assets/8dd45b04-06d9-4a6f-b777-4d9830991cf8)

## Why It's Good For The Game

The cape is very important for showing rank at a moment's notice and it
goes against standards to not wear one. Specism lame and stupid.

## Changelog

:cl: stiel0248, thgvr, cloudbreak
add: vox and kepori sprites for the azuilhauz
/:cl: