IPCs with the MMI brain setting now get an MMI brain (#4414)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

moves where the mutant brain is checked to before it's actually
generated so it does something

## Why It's Good For The Game

pref thing that didnt work now does work

## Changelog

:cl:
fix: IPCs with an MMI brain now get an MMI brain instead of it giving
them a positronic anyways. You are NOT a government drone.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Signed-off-by: Theos <theubernyan@gmail.com>