Bolide Armory matchup (Bonus new mattress for the Crying Sun) (#4355)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Matches up the cells and equipment in the Elated Bolide's armory to fit
the number of weapons and crew recieving said weapons. Prior, there was
an odd number of cells, and no civilian grade cells for the handguns,
which led to very confusing distribution. This changes the armory to
carry 2 military grade cells for each rifle (Same as the Crying Sun) and
2 civilian grade cells for each pistol.

Additionally adds 2 survival knives, as to equip each member of the away
team with a blade of some kind, bringing to total to 6 survival knives
and 2 machetes.

I also added 1 portable pump to the engineering supplies. 

On a really tiny note, I saw how the captain's quarters on the Bolide
used a double bed, tried it on the Crying Sun. Lo and behold it looked
better.

Pics below:
Armory Change
![Screenshot 2025-03-12
224624](https://github.com/user-attachments/assets/16974e3e-f587-4c17-943f-9cdfac1c616a)

_The mattress_
![Screenshot 2025-03-12
224633](https://github.com/user-attachments/assets/6f629ef7-e797-4f41-9a61-52ab4eeb4473)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Having played on the bolide a few times but never reaching a full crew,
the equipment issue never really arose, but now that I've seen a full
team and struggled to wrap my head around how the supplies are meant to
be distributed, I finally feel comfortable making the change.

Having a differentiation between military and civilian cells for the
handguns and rifles not only makes intended distribution easier to
discern but also keeps with the precedent set by all other instances of
the kalixcian handgun mapped in game.

With the knives change, it just seemed off for not every marine to have
a knife. The Marine Corp prides itself on its close combat abilities, so
leaving a marine out on a melee weapon (especially one as unremarkable
as a survival knife) felt wrong.

For the pump, it's just a convenience and there was a good unused spot
for it in storage. For a ship with an engineering specialty and the
supplies for making fortifications/ground sites, it seems quite thematic
and useful.

Lastly, the bed change. It looks nicer, that's it. Don't see how that's
an issue, especially in a room as unused as the Captain's Quarters.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added a portable pump to the Bolide
balance: Re-evaluated the bolide armory equipment
add: Added a new bed to the Crying Sun
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->