Quirk Rebalance +  Bonus Language Quirks (#4400)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
This also sorts quirks into individual files. also removes the infinite
pizza box. sorry I eated it all

junkie is still here along with smoker but is not in the screenshot.
these quirks are awful code though

Also adds four new language quirks, all exclusive with each other. This
means your character can freely pick one extra language.

Removes a lot of quirks that just sucked, didn't work, unused, etc. The
final list of quirks in the game is now this:

![image](https://github.com/user-attachments/assets/dc46f0a0-0f5a-4697-834f-c46fa2338e53)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
its not
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
add: New language quirks for the four main languages
del: Removed and rebalanced a lot of quirks
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->