Reduces E11 spread by half + Black Market price decrease (#4439)
## About The Pull Request

Reduces E11 spread from 60 to 30. Halved the dual wield and unwielded
spread as well.

Also decreases the max cost for the E11 in the black market.

## Why It's Good For The Game

There was talk about making the E11 at least semi-usable instead of
actually worthless. It is still the least accurate gun in the game by
several magnitudes, but you can probably hit someone within a single
tile of you.

## Changelog

:cl:
balance: Halved E11 spread and decreased the max cost for the E11 in the
black market
/:cl: