E11 hermits can now randomly spawn on sandworlds. (#4461)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR does a couple things:
- Makes the hermits spawn (5% taken out of melee hermit's chance rn, I'm
happy to tweak it cause it's taken out of my ass)
- Adds an eoehoma documents folder and places it in the E-11 ruin (so
the docs fit in the cabinet)
- Nerfs the E-11 drop chance from that hermit by 10%, so it's 40% now.
Still more than other hermits, but it's not like the gun is any good.
- Buffs the hermit's spread to reflect the current E-11 numbers
- Removes an accidental mission POI from the E-11 manufactory ruin
- Very minorly tweaks some descs
- Updates the hermit icons with current in-hands
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Those hermits are funny. Also, Apogee said they want more laser mobs...
so yeah...
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: E-11-wielding hermits started roaming the sand worlds.
tweak: Minor tweaks to the E-11 Manufactory ruin.
balance: Tweaked some numbers on the E-11 hermit.
imageadd: Updated the E-11 hermit sprites with current in-hands.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->