Language Icons in Runechat (#4469)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Language icons are now prepended to runechat messages when they would
also be shown in the chat log. The icons are scaled from their regular
icons, with antialiasing.


![image](https://github.com/user-attachments/assets/05f37161-0048-4e18-a760-6a0fe10f8e3e)
![image](https://github.com/user-attachments/assets/98d81704-ee8c-445a-b94c-c591b06504ca)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Makes it easier to tell what language someone, or yourself, is speaking
without needing to look at the chat log.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added language icons to runechat
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->