Microtarget with barrel is no longer bulky (#4508)
## About The Pull Request

With barrel it is now normal instead of bulky. Still small with barrel
off

Also decreases spread

## Why It's Good For The Game

You can put it in your holster I guess. Theres no way This Fucking Thing
is the same size as a full-sized rifle.

With the barrel off it is also more inaccurate than the E11. It is still
really bad, like Spitter levels of bad.

## Changelog

:cl:
balance: Microtarget with barrel is now Normal-sized instead of Bulky
balance: Decrease Microtarget spread
/:cl: