HOT-FIXES: Hot fix dynamic mission (#1697)
* Removes the Bubble-class (#4618)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Per #4488
> Colony ships simply aren't a niche we are pursuing anymore, ...

Also it's just.. outdated and unused in the first place
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Is removal ever good? Maybe not, but I don't anybody plans on using this
ship anytime soon, sadly.

![3dgifmaker52244](https://github.com/user-attachments/assets/d913c716-4523-4988-bb24-f26e663a7c8e)
###### (I spent more time on this gif than the removal)
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
del: Removed the Bubble-class
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

* Dynamic mission objectives can now be spawned in containers (#4616)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Title.

Dynamic mission landmarks can now be set to spawn their items inside
containers(or be marked a pre-existing one if already spawned in a
container)

Fixes the ruin mission on Wrecked Factory and Scrap Station.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Having your secret documents spawn on top of the locked safe is bad I
think.

## Changelog

:cl:
add: Dynamic mission objectives can now properly start inside containers
fix: Wrecked Factory and Scrapstation retrieval missions
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

* фикс наших карт

---------

Co-authored-by: Ical <86125936+Ical92@users.noreply.github.com>
Co-authored-by: Gristlebee <56049844+Gristlebee@users.noreply.github.com>