Day 1 Typhon tweaks (#4575)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
- Adds a bag of c4 - it's meant to be an assault ship, hard to get into
places without those, completely forgot about them.
- Adds a tank of firefighting foam - it starts with an adv. extinguisher
(cause of the Solaris that tends to set shit on fire), but I forgot to
give it a way to refill.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Changelog

:cl:
add: Added a bag of c4 and a firefighting foam tank to the Typhon-class.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->