TG modernization of event code (#4545)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
handful of tg station code improvments to events that act as a prereq
for some new events i was cooking up, however since they are still
disabled none of this is player facing.

https://github.com/tgstation/tgstation/pull/72998
https://github.com/tgstation/tgstation/pull/73801
https://github.com/tgstation/tgstation/pull/83424
any many more... thank you tgstation

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
im cookin up some stuff for events.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: FalloutFalcon, too many TG station contributers to name
refactor:  buncha refactors to events and holidays from tgstation
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->