Makes the indie viper-23 double action. (#4630)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Before it was single-action, different from the military version which
is double action, now they're both the same, with the indie version
having a bit more spread because it's shorter.

Apogee's request. Having the same gun work completely differently
between the military and civilian versions is weird

:cl:
balance: The Viper-23 is now double action same as the military variant,
with a bit more spread to compensate.

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->