ECM-6 purchasable in cargo + CMM-BARD to CLIP-BARD (#4536)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
* Changes mentions of CMM-BARD to CLIP-BARD
* Reflavors the CM-1 to the ECM-6 and makes it on-par with the standard
/kalix laser gun
  * Also added it to cargo for 1500 credits
* If you really wanna make an Elzu or Vox veteran minuteman who lived
through the entirety of CLIP or something there's an ECM-1 subtype that
uses old cells and only fires semi-auto
 * Does some slight rewriting of the descs of CLIP guns

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

this has been rotting for a year now and the minutemen crave positonics

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: The ECM-6 is now in cargo, available at the request of CLIP vessels
spellcheck: CMM-BARD to CLIP-BARD
spellcheck: Changed the description of some CLIP weapons
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->