Replaces the "oxygen" cans filled with plasma with oxygen cans on cave base. (#4764)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
says in the title.


## Why It's Good For The Game
Gotcha mapping is not good. There is not hint of plasma in the ruin
anyways.

## Changelog

:cl:

fix: removes the fake oxygen cans on cave base and replaces them with
actual oxygen.

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->