Balance pass over gibonite (#4698)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
gibonite detonated prematurely does very little (means if someone drills
to fast if wont murder them)
nerfs all the explosions of gibonite but buffs the item light range so
its better for mining while not nuking a ship if someone accidentally
hits one
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
alternative to fully removing them, though i dont think this is the only
pr required, i think they should be reflavored into some sort of
unstable plasma or sulfer that can be refined for a good chunk of it.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: New round tip to let people know you can defuse gibonite with a
multi-tool.
balance: Gibonite explosions are weaker, premature explosions do
virtually nothing.
balance: The light explosion range of the Gibonite item is a little
stronger.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->