Events! (#4546)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
Revival of #3070
Requires #4545

Ping me if you have a cool idea for a new event.

This re-enables events but first reworks, removes, and adds missions.

The current list of automatic events you would see in a round is as
follows.

**Random Pet!**
spawns a random pet hidden inside a random crate of a ship
**Rat Infestation!**
spawns a few rats on a random ship!
**Lottery!**
A ship wins a bunch of creds, announced for all ships to hear.
**Random Cargo!**
A ship is struct with a drop pod containing a random cargo pack. They
are given ample warning and the damage is minimal
**Spammer!**
fax machines as spammed with junk mail!

Admin Spawn (Mostly because WIP)
**Frontiersmen** (Outpost places a bounty on them, or they are tasked to
hyjack the outpost)
**The Salesman** (A little guy that shows up at your local outpost and
sells you stuff
**Ion storm** (I was annoying to remove, broken but has stuff refrencing
it rn)
**Random Sentience**(I think its cute but needs work)
**Harmless ion clouds**(doenst work as startlight is config disabled)

Anything not mentioned was CUT out of the code with no mercy.
This also removed a bit of random antag cruft. only thing player facing
was the pref menu which is now not full of fake antags.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
events cool, adds spice to the rounds for ships that dont have like an
admin event being ran.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
add: Readded events, mostly replacing them with new ones
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>