Fixes grav anomaly core detonation (#4783)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Detonating anomaly cores has an opportunity cost of ~3000 credits so you
will NEVER see these but the grav anomaly now properly deletes itself
and attempts to do the mega ultra bone break before the smaller one
which allows it to work

## Why It's Good For The Game

Working As Intended

## Changelog

:cl:
fix: detonating a grav anomaly core can now break bones as intended and
also deletes the core as intended
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Signed-off-by: Theos <theubernyan@gmail.com>