gunslinger spread reduction buff is no longer randomized (#4787)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

The application of gunslinger's spread increase/reduction was modified
so it wouldn't GUARANTEE they miss with bad weapons (the 30 degree
forced spread increase has arrived) but this also impacted the buff on
revolvers making it less reliable and significantly less effective

## Why It's Good For The Game

I wanna shoot straight with my handcannon 💀 💀 💀 

## Changelog

:cl:
fix: the spread reduction for gunslinger provided to revolvers is no
randomized when applied
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->