Gets handrails out of the way & adds handrails to Riggs (#4601)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Moves handrails out from the middle of cargo bay to more appropriate
spots against the walls or railings on the Raleigh, Scarab, and Elated
Bolide. Adds handrails to the Riggs, which has somehow not had any all
this time.

## Changelog

:cl:
add: Added handrails to Riggs
fix: Moved handrails out of the way on Raleigh, Scarab, Elated Bolide
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: thearbiber <135081923+thearbiber@users.noreply.github.com>