Allows Binoculars to be worn around the neck (#4830)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

This PR allows binoculars to fit in the neck slot of a character. This
PR also adds a mob overlay, courtesy of sprites from @meemofcourse

![image](https://github.com/user-attachments/assets/9ae4a508-ec3d-4457-bfcd-593f4ae1aad6)

![image](https://github.com/user-attachments/assets/8b2078c3-624d-4044-8870-9e2d8efc57a5)

I thought it would add a tiny bit of flavor to be able to strap a pair
of binoculars around a character's neck. Ultimately, not a very
interesting change, however.

:cl: Meemofcourse
add: Added the ability to wear binoculars around one's neck
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->