Fixes fancy chat failing to load (#4811)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
prevents this by porting
https://github.com/tgstation/tgstation/pull/70386

![442362872-8795b109-f69a-4a3d-ab62-eceb67a2add6](https://github.com/user-attachments/assets/f1ba8d03-1baa-40c5-b6b7-1f119f3d76f3)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I hate this solution as even tho they said it minutes is still limited.
why even have a seconds topic limiter if its this easy to bypass.
but.
fixes chat so. yipeeeeee
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: fancy chat should now consistently load. Please ping @_fallcon if
it doesn't.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: Theos <theubernyan@gmail.com>