Better cargo pack managment (#4419)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
refactors packs to be stored in a /datum/cargo_market, my goal is to
merge the blackmarket, outpost packs, and custom markets(neat for
adminbus) into the same system (with different subtypes)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Gives us the ablility to make cooler custom markets for admin bus or new
outpost types or something, will eventually merge the behavoir of
blackmarket into this datum so we have access to black market features
in regular markets vice versa.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
add: Manifest for cargo purchases are formatted a lot better
refactor: Refactored cargo packs markets to be managed by a datum
instead of the outpost. This also should result in orders being grouped
better.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Ical <wolfsgamingtips@gmail.com>