Adding the Mercurial Quasar (#4845)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds in the Mercurial Quasar-class Rapid Reconnaissance Vessel, a small
PGF Vessel which maintains a combination of PGFMC and Navy crew.

The Mercurial Quasar was originally designed to support larger fleet
formations as a forward scouting vessel and more specifically as an
ambush clearing vessel. When a hostile presence is expected, but not yet
identified, sensors won't cut it and the drones are coming up dry,
Mercurial Quasars are deployed to find the way.

Heavily armed for it's size with an impressive point defense suite on
it's nose and a pair of lasers on it's tail, the Mercurial Quasar can
hold it's own in the face of oncoming fire. With equally powerful
propulsion, the Mercurial Quasar has the power it needs to get out of a
bad situation in one piece.

Now in service of the PGF's frontier expansion initiative, Mercurial
Quasars are being given new purpose. On board armaments are more than
enough to dissuade small fry pirates and the vessel's rapid deployment
capabilities, along with a highly cross trained crew, makes the vessel
extremely flexible for common operations on the frontier.

Crew Compliment:
x1 Captain
x1 Navy Engineer
x1 Navy Corpsman
x1 Marine Sergeant
x2 Marines

Images:

![image](https://github.com/user-attachments/assets/717c87be-7973-4b92-ad9c-9e88da8e9839)

![image](https://github.com/user-attachments/assets/63f5f156-8f16-4cef-9107-26776016c857)
![Screenshot 2025-05-23
165444](https://github.com/user-attachments/assets/d54e54f3-57bb-4951-98f7-f90e5aed244b)
![Screenshot 2025-05-23
165459](https://github.com/user-attachments/assets/99316e69-c268-481f-942a-1310282614f5)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
Currently, if you want the full PGF experience with marines and navy
crew together, you're forced to use a much larger ship in the form of
the Crying Sun and Elated Bolide, which isn't always feasible for
smaller groups of players, or is out of the comfort zone of newer
captains. This ship helps fill that currently unfilled niche in the PGF
roster.
## Changelog

:cl:
add: Mercurial Quasar-class Rapid Reconnaisance Vessel
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: DIB-DOG <DaneOfThieves@gmail.com>
Co-authored-by: Changelogs <action@github.com>