Adds missing inhands for rollie-type cigarettes (#4930)
## About The Pull Request
A missing icon resulted in rollies not having an inhand for... about 4
years or so. This PR adds inhand sprites that are functionally identical
(aside from a pixel of green) to the cigarette inhands for rollies.


## Why It's Good For The Game
I want to be able to smoke a blunt and take it out of my mouth and hold
it and have people see that i am holding a blunt. my massive spliff. A
fat Dart. A Cungadero.


![topshelfzaza](https://github.com/user-attachments/assets/3a3201a8-d9b4-4fe4-9eb6-02a5649e6a3c)


## Changelog

:cl:
add: Rollies now have an inhand sprite.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->