Pouch whitelist contents update -- bodycameras and welding goggles (#4976)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds the ability to hold welding goggles in technician's
pouches, and it also allows one to hold and view bodycameras from within
command, technician's, and basic pouches.

## Why It's Good For The Game

It makes sense to be able to hold a small form of eye protection in a
pouch. It's also nice to be able to have more space to hold a body
camera in, so that you don't have to sacrifice more vital space for one.

## Changelog

:cl:
balance: welding goggles now fit in technician's pouches
balance: bodycameras now fit in command and technician's pouches
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Theos <theubernyan@gmail.com>