PGF-Ship-touchups-+-Cthonian-overhaul (#4943)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR does a number of housekeeping updates to the PGF ship lineup,
mostly reorganization or the addition of previously missing uniform
items. Examples include the addition of gloves to the Elated Bolide's
uniform locker, the addition of medical gloves to the Crying Sun's
doctor locker, the addition of hand labelers to the rest of the PGF cryo
rooms, adding missing light switches, allowing the PGF to buy it's
factional marine armor and many more such minor qol changes.

The meat of the PR comes with the overhaul to the Woeful Cthonian.
Previously serving as a generalist craft, the Woeful Cthonian has been
shifted to a medical and telecommunications focus, in addition to it's
original light reconnaissance duties. Accordingly, the ship has been
adjusted to accommodate these new roles with an expanded medical bay and
the addition of a telecommunications room in the re-modeled engineering
bay.

Images:
New Cthonian

![image](https://github.com/user-attachments/assets/9f9bb02e-a44f-40f3-b8ec-19e167c1aac9)
With areas

![image](https://github.com/user-attachments/assets/385e920a-c579-495e-9c10-c16ea93e5a06)
Old Cthonian

![image](https://github.com/user-attachments/assets/5829c901-aed2-40b7-8ffd-5424b6d029b1)
Armor mishap

![image](https://github.com/user-attachments/assets/6b7bfe11-21e2-4e2b-8dcb-345aff61385c)
Fixed it

![image](https://github.com/user-attachments/assets/69576d6b-a133-4732-9582-1eb622cd584c)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
WIth the addition of the Mercurial Quasar to the PGF vessel lineup,
which more fluidly serves the role of a small crew recon vessel, the
Woeful Cthonian has felt oddly left out. Without dedicated combat roles
such as marines, it often feels awkward sending sailors in to storm the
beaches.

By providing the Cthonian with the ability to specialize in medical work
and telecommunications connection, it's mission and purpose become much
clearer, offering direction to what are often confused or aimless crews.

Secondly, this modification to the Cthonian helps to fill the existing
gap in the PGF roster by providing the faction with both a medical and
telecomms ship in one.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added a CMO PGF doctor variant
balance: Overhauled the Woeful Cthonian
fix: fixed missing items on three PGF ships
fix: PGF marine armor crates and PGF navy coat crates now have unique
paths
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: DIB-DOG <DaneOfThieves@gmail.com>
Co-authored-by: Changelogs <action@github.com>