Ion Rifle Touch Up (#5016)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Reduces the cost of the ion rifle from 10k to 3500

Ion rifle is now bulky instead of huge, and now has a roughly six second
cooldown between shots. An overlay has been added to indicate when it's
on cooldown.

## Why It's Good For The Game

With #5002, EMPs should no longer be horrifically oppressive to
augged/mechanical players, so I think it would be interesting to have
the ion rifle not cost a ludicrous amount since that's no longer a
concern. It should open up some new combat options without being unfun
to play against now.

The cooldown on the rifle is to prevent chain stuns or spam, and I'm
happy to adjust it as needed.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
balance: Ion rifle price reduced from 10k to 3500
balance: Ion rifle has a six second cooldown between shots.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->