New Ruin: Nomad's Stop (#5034)
## About The Pull Request
![StrongDMM-2025-07-04 23 45
58](https://github.com/user-attachments/assets/d150bad4-d03b-46ef-b4c3-2125844180a7)
![StrongDMM-2025-07-04 23 46
01](https://github.com/user-attachments/assets/733f2add-64f3-41ff-be95-c0d75236860b)

![dreamseeker_WbzzlrpLUG](https://github.com/user-attachments/assets/9437ca58-d2ab-4bbd-8089-b64516a217be)

![dreamseeker_x53gjNNzDH](https://github.com/user-attachments/assets/3ac11531-3d64-4851-b217-cc90399c077e)

It's a few buildings that I filled with random crap and built a few
contraptions for. I'm not entirely sold on it but it's far along enough
I'd like some actual reviews.
Also I am tired of looking on it I'd like to move onto a new project.


## Why It's Good For The Game
new ruin
## Changelog

:cl:
add: A small hermit trade post has been detected on the surface of your
nearest sand planet
/:cl:

---------

Co-authored-by: thearbiber <135081923+thearbiber@users.noreply.github.com>