Heavy explosions deal less damage to items (#5040)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Heavy bomb damage range to items is now from 90 to 150, down from 90 to
250.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Explosions are the main source of item shredding/destruction and feels
terrible currently. Item integrity by default is 200, meaning with the
original values you effectively have a 1/3 chance to have items be
immediately destroyed by a heavy explosion. It'll take at least two
explosions now generally to fully destroy an item. I think it's more
engaging to have a chance to repair your broken gear rather than deal
with instant destruction, and if you do happen to get hit by a second
explosion at that point it's probably on you.

## Changelog

:cl:
balance: Max damage from heavy explosions to items is now 150, from 250
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->