Breach Mining (#5057)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
Changes how drilling works.

Drills now spawn "seismic disturbances" that spawn the monsters. They
stay in place for the duration of the "wave". After a certain amount of
waves you get a short break that is EXPLICITLY NOTIFIED

![dreamseeker_euQbH5FbEw](https://github.com/user-attachments/assets/a3bebab0-62c4-4b88-b63c-ef1818e1d075)

![dreamseeker_k03O0zdMKd](https://github.com/user-attachments/assets/95c8ce1f-7013-4cd9-a2aa-10c176fbc37e)

:cl:
add: Drills now spawn a spawner, which in turn spawns the monsters. It's
an innovative system really.
fix: ACTUALLY FIXED DRILL TURF PLACEMENT THIS TIME
/:cl: