Thermal Eyepatch Price Increase (#5089)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! --> Increases the price for the thermal eyepatch, an
item that is rather powerful, all things considered. It also fixes a
typo on the Ablative Trenchcoat listing.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->Being able to see everything on the other side of a
wall is not something you should be able to have for below 2000. The
original minimum was 1000 with a max of 3000. More often than not, one
would see it on the low end. Its is now a minimum of 2500 and a max of
3500.

## Changelog

:cl:
balance: Price of the Thermal Eyepatch has been increased
fix: Taught the black market product lister how to spell trenchcoat.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->