Removes the chance for bulbs and light tubes to start out broken in ship and outpost areas (#5169)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Sometimes when a ship is spawned in, some of its lights start out of the
gate being broken. Ultimately only a minor nuisance, but a nuisance
nevertheless. This PR should toggle that behavior from select areas.

A flag has been added to areas labeled 'NO_RANDOM_LIGHT_BREAKAGE' which
determines if an area will allow a light tube or a light bulb to
randomly break during late initialization. This flag is added to ship
areas and outpost areas by default.

## Why It's Good For The Game

While it does give some busywork, I feel like it may be nicer to have
this behavior be gone. I'm open to other opinions on this, however.

The reason why the flag was added to both ship areas and outpost areas
is under the idea that both of these areas are supposed to be relatively
well maintained (as opposed to being dilapidated, like how a ruin may
be)

## Changelog

:cl:
add: Added a flag to areas to toggle light breakage behavior during late
initialization
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->