Allows paper to be put in pill bottles (#4889)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

IRL medications are usualy sold with a little leaflet containing a bunch
of info such as: the producer, dosage and how often to take it, the
effects and the risks, etc...

## Why is it good for the game

I think it would be flavorful.

## Changelog

:cl:
add: You can now put paperslips in a pill bottle. Paper slips are
craftable by slicing paper with a knife.
:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: ritorizo <41831966+ritorizo@users.noreply.github.com>
Co-authored-by: ritorizo <ritorizo@localhost>
Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>