Make items dipable in barrels. (#5235)
## About The Pull Request
It's possible to dip items in a glass or beaker for various purposes
(making concrete with cement and sand, turning iodine into bromine with
a scanning module...).
This pr extends this behavior to barrels.  

Also makes it so you can only see the volume of the reagents in a barrel
if it's open.

## Why It's Good For The Game
Barrels are essential for large chem mixes and this pr makes working
with them less tedious.

## Changelog
:cl:
add: It's now possible to dip items in barrels.
fix: You can no longer see the volume of the content of a closed barrel.
:cl:

Co-authored-by: ritorizo <ritorizo@localhost>