Add backpacks to the clothing vendor (#5244)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Adds 4 backpacks, satchels and duffel bags to the ClothesMate at the
price of 50 credits (same as in cargo).

You shouldn't need your captain to buy a bag.

:cl:
add: Backpacks, satchels and duffel bags are now in the ClothesMate for
50 credits
:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: ritorizo <ritorizo@localhost>